+++ libs/gesha-core/src/io/mod.rs
@@ -3,6 +3,3 @@ pub use reader::Reader;
 
 mod writer;
 pub use writer::Writer;
-
-mod diff;
-pub use diff::detect_diff;
