+++ libs/gesha-core/src/testing/test_runner.rs
@@ -1,6 +1,6 @@
 use crate::conversions::Generator;
-use crate::io::{detect_diff, Writer};
-use crate::testing::{run_parallel, TestCase, TestCaseIndex, TestDefinition};
+use crate::io::Writer;
+use crate::testing::{detect_diff, run_parallel, TestCase, TestCaseIndex, TestDefinition};
 use crate::{Error, ErrorTheme, Result};
 use std::fmt::Debug;
 use tracing::{info, instrument};
